회원 자세히 보기 view 구현

+++ momo_project/src/main/java/com/momo/admin/AdminService.java
@@ -45,7 +45,7 @@ public class AdminService {
 	}
 	
 	//해당 회원 조회
-	public Member findMemberByMemNum(int memId) {
+	public Member findMemberByMemNum(long memId) {
 		Member member = adminMemberRepository.findById(memId).get();
 		
 		return member;
